// packages/figma-parser/src/full-figma-types.d.ts
export declare type StyleType = 'FILL' | 'TEXT' | 'EFFECT' | 'GRID';
// ...
export declare type BlendMode = 'PASS_THROUGH' /** (Only applicable to objects with children) */ | 'NORMAL'
  /** Darken: */
  | 'DARKEN' | 'MULTIPLY' | 'LINEAR_BURN' | 'COLOR_BURN'
  /** Lighten: */
  | 'LIGHTEN' | 'SCREEN' | 'LINEAR_DODGE' | 'COLOR_DODGE'
  /** Contrast: */
  | 'OVERLAY' | 'SOFT_LIGHT' | 'HARD_LIGHT'
  /** Inversion: */
  | 'DIFFERENCE' | 'EXCLUSION'
  /** Component: */
  | 'HUE' | 'SATURATION' | 'COLOR' | 'LUMINOSITY';
// ...
export declare type Node = Document | Canvas | Frame | Group | Vector | BooleanOperation | Star | Line | Ellipse | RegularPolygon | Rectangle | Text | Slice | Component | ComponentSet | Instance;
/** Node Properties */
/** The root node */
export interface Document extends Global {
  readonly type: 'DOCUMENT';
  /** An array of canvases attached to the document */
  readonly children: ReadonlyArray<Node>;
}
/** Represents a single page */
export interface Canvas extends Global {
  readonly type: 'CANVAS';
  /** An array of top level layers on the canvas */
  readonly children: ReadonlyArray<Node>;
  /** Background color of the canvas */
  readonly backgroundColor: Color;
  /** Node ID that corresponds to the start frame for prototypes */
  readonly prototypeStartNodeID: string | null;
  /** An array of export settings representing images to export from the canvas */
  readonly exportSettings?: ReadonlyArray<ExportSetting>;
}
// ...
export interface Frame extends FrameBase {
  readonly type: 'FRAME';
}
/** A logical grouping of nodes */
export interface Group extends FrameBase {
  readonly type: 'GROUP';
}
export interface VectorBase extends Global {
  /**
   * An array of export settings representing images to export from node
   * @default []
   */
  readonly exportSettings?: ReadonlyArray<ExportSetting>;
  /**
   * How this node blends with nodes behind it in the scene
   * (see blend mode section for more details)
   */
  readonly blendMode: BlendMode;
  /**
   * Keep height and width constrained to same ratio
   * @default false
   */
  readonly preserveRatio?: boolean;
  /**
   * Horizontal and vertical layout constraints for node
   */
  readonly constraints: LayoutConstraint;
  /**
   * Node ID of node to transition to in prototyping
   * @default null
   */
  readonly transitionNodeID?: string | null;
  /**
   * The duration of the prototyping transition on this node (in milliseconds)
   * @default null
   */
  readonly transitionDuration?: number | null;
  /**
   * The easing curve used in the prototyping transition on this node
   * @default null
   */
  readonly transitionEasing?: EasingType | null;
  /**
   * Opacity of the node
   * @default 1
   */
  readonly opacity?: number;
  /** Bounding box of the node in absolute space coordinates */
  readonly absoluteBoundingBox: Rect;
  /**
   * Width and height of element. This is different from the width and height
   * of the bounding box in that the absolute bounding box represents the
   * element after scaling and rotation. Only present if geometry=paths
   * is passed
   */
  readonly size?: Vector2;
  /**
   * The top two rows of a matrix that represents the 2D transform of this
   * node relative to its parent. The bottom row of the matrix is implicitly
   * always (0, 0, 1). Use to transform coordinates in geometry.
   * Only present if geometry=paths is passed
   */
  readonly relativeTransform?: Transform;
  /**
   * An array of effects attached to this node
   * (see effects sectionfor more details)
   * @default []
   */
  readonly effects: ReadonlyArray<Effect>;
  /**
   * Does this node mask sibling nodes in front of it?
   * @default false
   */
  readonly isMask?: boolean;
  /**
   * An array of fill paints applied to the node
   * @default []
   */
  readonly fills: ReadonlyArray<Paint>;
  /**
   * Only specified if parameter geometry=paths is used. An array of paths
   * representing the object fill
   */
  readonly fillGeometry?: ReadonlyArray<Path>;
  /**
   * An array of stroke paints applied to the node
   * @default []
   */
  readonly strokes: ReadonlyArray<Paint>;
  /** The weight of strokes on the node */
  readonly strokeWeight: number;
  /**
   * Only specified if parameter geometry=paths is used. An array of paths
   * representing the object stroke
   */
  readonly strokeGeometry?: ReadonlyArray<Path>;
  /**
   * Where stroke is drawn relative to the vector outline as a string enum
   * "INSIDE": draw stroke inside the shape boundary
   * "OUTSIDE": draw stroke outside the shape boundary
   * "CENTER": draw stroke centered along the shape boundary
   */
  readonly strokeAlign: 'INSIDE' | 'OUTSIDE' | 'CENTER';
  /**
   * Styles this node uses from the global `styles`
   */
  readonly styles?: StylesObject;
}
/** A vector network, consisting of vertices and edges */
export interface Vector extends VectorBase {
  readonly type: 'VECTOR';
}
/** A group that has a boolean operation applied to it */
export interface BooleanOperation extends VectorBase {
  readonly type: 'BOOLEAN_OPERATION';
  /**
   * A string enum with value of "UNION", "INTERSECT", "SUBTRACT", or "EXCLUDE"
   * indicating the type of boolean operation applied
   */
  readonly booleanOperation: 'UNION' | 'INTERSECT' | 'SUBTRACT' | 'EXCLUDE';
  /** An array of nodes that are being boolean operated on */
  readonly children: ReadonlyArray<Node>;
}
/** A regular star shape */
export interface Star extends VectorBase {
  readonly type: 'STAR';
}
/** A straight line */
export interface Line extends VectorBase {
  readonly type: 'LINE';
}
/** An ellipse */
export interface Ellipse extends VectorBase {
  readonly type: 'ELLIPSE';
}
/** A regular n-sided polygon */
export interface RegularPolygon extends VectorBase {
  readonly type: 'REGULAR_POLYGON';
}
// ...
export interface Text extends VectorBase {
  readonly type: TextType;
  /** Text contained within text box */
  readonly characters: string;
  /**
   * Style of text including font family and weight (see type style
   * section for more information)
   */
  readonly style: TypeStyle;
  /**
   * Array with same number of elements as characeters in text 'box' |    * each element is a reference to the styleOverrideTable defined
   * below and maps to the corresponding character in the characters
   * field. Elements with value 0 have the default type style
   */
  readonly characterStyleOverrides: ReadonlyArray<number>;
  /** Map from ID to TypeStyle for looking up style overrides */
  readonly styleOverrideTable: {
    readonly [index: number]: TypeStyle;
  };
}
// ...
export interface Component extends FrameBase {
  readonly type: 'COMPONENT';
}
/** A node that can have multiple component variations */
export interface ComponentSet extends FrameBase {
  readonly type: 'COMPONENT_SET';
}
/**
 * An instance of a component, changes to the component result in the same
 * changes applied to the instance
 */
export interface Instance extends FrameBase {
  readonly type: 'INSTANCE';
  /**
   * ID of component that this instance came from, refers to components
   * table (see endpoints section below)
   */
  readonly componentId: string;
}
/** An RGBA color */
export interface Color {
  /** Red channel value, between 0 and 1 */
  readonly r: number;
  /** Green channel value, between 0 and 1 */
  readonly g: number;
  /** Blue channel value, between 0 and 1 */
  readonly b: number;
  /** Alpha channel value, between 0 and 1 */
  readonly a: number;
}
// ...
export interface Effect {
  /** Type of effect as a string enum */
  readonly type: 'INNER_SHADOW' | 'DROP_SHADOW' | 'LAYER_BLUR' | 'BACKGROUND_BLUR';
  /** Is the effect active? */
  readonly visible: boolean;
  /** Radius of the blur effect (applies to shadows as well) */
  readonly radius: number;
  readonly color?: Color;
  readonly blendMode?: BlendMode;
  readonly offset?: Vector2;
  readonly spread?: number;
}
/** A solid color, gradient, or image texture that can be applied as fills or strokes */
export interface Paint {
  /** Type of paint as a string enum */
  readonly type: PaintType;
  /**
   * Is the paint enabled?
   * @default true
   */
  readonly visible?: boolean;
  /**
   * Overall opacity of paint (colors within the paint can also have opacity
   * values which would blend with this)
   * @default 1
   */
  readonly opacity?: number;
  /** Solid color of the paint */
  readonly color?: Color;
  /**
   * How this node blends with nodes behind it in the scene
   * (see blend mode section for more details)
   */
  readonly blendMode: BlendMode;
  /**
   * This field contains three vectors, each of which are a position in
   * normalized object space (normalized object space is if the top left
   * corner of the bounding box of the object is (0, 0) and the bottom
   * right is (1,1)). The first position corresponds to the start of the
   * gradient (value 0 for the purposes of calculating gradient stops),
   * the second position is the end of the gradient (value 1), and the
   * third handle position determines the width of the gradient (only
   * relevant for non-linear gradients).
   *
   */
  readonly gradientHandlePositions?: ReadonlyArray<Vector2>;
  /**
   * Positions of key points along the gradient axis with the colors
   * anchored there. Colors along the gradient are interpolated smoothly
   * between neighboring gradient stops.
   */
  readonly gradientStops?: ReadonlyArray<ColorStop>;
  /** Image scaling mode */
  readonly scaleMode?: ScaleMode;
  /**
   * Affine transform applied to the image, only present if scaleMode is `STRETCH`
   */
  readonly imageTransform?: Transform;
  /**
   * Amount image is scaled by in tiling, only present if scaleMode is `TILE`
   */
  readonly scalingFactor?: number;
  /**
   * A reference to an image embedded in the file. To download the image using this reference,
   * use the GET file images endpoint to retrieve the mapping from image references to image URLs
   */
  readonly imageRef?: string;
  /**
   * A reference to the GIF embedded in this node, if the image is a GIF.
   * To download the image using this reference,
   * use the GET file images endpoint to retrieve the mapping from image references to image URLs
   */
  readonly gifRef?: string;
}
export interface Path {
  /** A sequence of path commands in SVG notation */
  readonly path: string;
  /** Winding rule for the path */
  readonly windingRule: 'EVENODD' | 'NONZERO';
}
export declare type Transform = ReadonlyArray<ReadonlyArray<number>>;
/** A 2d vector */
export interface Vector2 {
  /** X coordinate of the vector */
  readonly x: number;
  /** Y coordinate of the vector */
  readonly y: number;
}
/** A position color pair representing a gradient stop */
export interface ColorStop {
  /** Value between 0 and 1 representing position along gradient axis */
  readonly position: number;
  /** Color attached to corresponding position */
  readonly color: Color;
}
/** Metadata for character formatting */
export interface TypeStyle {
  /** Font family of text (standard name) */
  readonly fontFamily: string;
  /** PostScript font name */
  readonly fontPostScriptName: string;
  /** Space between paragraphs in px, 0 if not present */
  readonly paragraphSpacing?: number;
  /** Paragraph indentation in px, 0 if not present */
  readonly paragraphIndent?: number;
  /** Is text italicized? */
  readonly italic?: boolean;
  /** Numeric font weight */
  readonly fontWeight: 100 | 200 | 300 | 400 | 500 | 600 | 700 | 800 | 900;
  /** Font size in px */
  readonly fontSize: number;
  /** Horizontal text alignment as string enum */
  readonly textAlignHorizontal: 'LEFT' | 'RIGHT' | 'CENTER' | 'JUSTIFIED';
  /** Vertical text alignment as string enum */
  readonly textAlignVertical: 'TOP' | 'CENTER' | 'BOTTOM';
  /** Space between characters in px */
  readonly letterSpacing: number;
  /** Paints applied to characters */
  readonly fills?: ReadonlyArray<Paint>;
  /** Line height in px */
  readonly lineHeightPx: number;
  /** Line height as a percentage of normal line height */
  readonly lineHeightPercent: number;
  /** The unit of the line height value specified by the user. */
  readonly lineHeightUnit: 'PIXELS' | 'FONT_SIZE_%' | 'INTRINSIC_%';
  /** Text casing applied to the node, default is the original casing */
  readonly textCase?: 'UPPER' | 'LOWER' | 'TITLE';
  /** Text decoration applied to the node, default is none */
  readonly textDecoration?: 'STRIKETHROUGH' | 'UNDERLINE';
  /** Line height as a percentage of the font size. Only returned when lineHeightPercent is not 100. */
  readonly lineHeightPercentFontSize?: number;
}
// ...
interface SharedElement extends ComponentMetadata {
  /** The unique identifier of the figma file which contains the element */
  readonly file_key: string;
  /** Id of the component node within the figma file */
  readonly node_id: string;
  /** URL link to the element's thumbnail image */
  readonly thumbnail_url: string;
  /** The UTC ISO 8601 time at which the element was created */
  readonly created_at: string;
  /** The UTC ISO 8601 time at which the element was updated */
  readonly updated_at: string;
  /** The user who last updated the element */
  readonly user: User;
}
// ...
export interface FullStyleMetadata extends SharedElement {
  /** The type of style */
  readonly style_type: StyleType;
  /** A user specified order number by which the style can be sorted */
  readonly sort_position: string;
}